requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
webdriver-manager==4.0.1
openai==1.3.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
schedule==1.2.0
qrcode==7.4.2
pillow==10.1.0
requests==2.31.0
